+++ app/(main)/page.tsx
@@ -1,7 +1,5 @@
 import { redirect } from "next/navigation";
 
 export default function Home() {
-  console.log("Home");
-
   redirect("/blog");
 }
