--- app/components/BlogButton.tsx
@@ -1,5 +1,0 @@
-import { NavButton } from "./navButton";
-
-export const BlogButton = () => {
-  return <NavButton label="Blog" href="/blog" />;
-};
