--- app/components/DraftButton.tsx
@@ -1,8 +1,0 @@
-import { NavButton } from "./navButton";
-import { auth } from "@/auth";
-
-export const DraftButton = async () => {
-  const session = await auth();
-
-  return session?.user && <NavButton label="Draft" href="/md" />;
-};
