+++ app/components/navButton.tsx
@@ -14,8 +14,7 @@ export const NavButton = ({
     href={href ?? ""}
     className={twMerge(
       "px-8 py-3 w-[150px] text-center  bg-transparent hover:bg-indigo-700/50",
-      isActive &&
-        "underline decoration-wavy decoration-2 underline-offset-8 decoration-white"
+      isActive && "bg-indigo-700/50"
     )}
   >
     {label}
